Workaround to avoid the led doesn't on sometimes
In my opnion, the correct way to enable pwm is:

    pwm_config(pwm_dev, duty_us, period_us);
    pwm_enable(pwm_dev);

And the correct way to disable pwm is:

    pwm_config(pwm_dev, 0, period_us);
    pwm_disable(pwm_dev);

However, when I operate the pwm device like the above, the led doesn't
on sometimes. It seems that `pwm_config` doesn't take effect.  I believe
this is due to some bugs of pm8xxx-pwm.

In this moment, I just change the order of `pwm_config` and `pwm_enable`
to avoid these problem happen.

Category: led
Bug: None

Change-Id: I6fa407354d70e7b39c1bcf503732fc74d8e088bf