CR# 421559: Change Apps MSM8930 PM8038 SMPS 3 sleep set mode request to 'auto'
Description: The modem processor may crash due to the SMPS 3 regulator
voltage dropping too low when the modem processor is heavily loaded
and the application processor is asleep. This happens because the
application processor votes for the SMPS 3 regulator mode to be PFM in
its sleep set and the modem processor makes no SMPS 3 request.
Therefore, when the application processor is asleep, the RPM switches
SMPS 3 into PFM mode. After that point, if the modem processor draws
more current than PFM mode can supply, the SMPS 3 voltage droops.

Change Description: The application processor sleep set mode request
was changed to 'auto'. This ensures that SMPS 3 is kept in auto mode
all of the time regardless of whether or not the application processor
is active or asleep.

Category: PM
Bug: None

Change-Id: I258cfe73e7a1e9fdd6ef4dec56ceaf934b8dfe8e